SlabInfo array initialization added

+++ kernel/src/memory_management/mod.rs
@@ -2,7 +2,7 @@ mod physical_memory_manager;
 mod virtual_memory_manager;
 
 /// 4KB
-const PAGE_SIZE: u64 = 4096;
+const PAGE_SIZE: usize = 4096;
 
 /// Inits Physical Memory Manager and Virtual Memory Manager
 pub fn init(boot_info: &bootloader_api::BootInfo) {
